requirements.txt
torch
torchvision
timm
transformers
albumentations
opencv-python
rich
numpy
pyyaml
tqdm
matplotlib
